Se agregó librería de tabs de react para el manejo de información dentro de la misma pantalla.
