+++ my-app/package.json
@@ -10,6 +10,7 @@
         "react-dom": "^16.9.0",
         "react-router-dom": "^5.0.1",
         "react-scripts": "3.1.1",
+        "react-tabs": "^3.0.0",
         "react-with-firebase-auth": "^1.0.4"
     },
     "scripts": {
