+++ my-app/src/components/Buttons/EntryButton/style.css
@@ -2,4 +2,5 @@
     background-color: #00BCD4 !important;
     height: 3rem;
     margin-top: 1rem;
+    margin: 1rem;
 }
