+++ my-app/src/components/Buttons/flatButton/index.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import './style.css';
 import Button from '@material-ui/core/Button';
 
-const FlatButton = ({text, color}) => {
+const FlatButton = ({text, color, onClick}) => {
     return (
-        <Button style={{Color:color}} className="flat-button">
+        <Button style={{Color:color}} className="flat-button" onClick={onClick}>
             {text}
         </Button>
     )
