+++ my-app/src/components/modal/style.css
@@ -11,7 +11,7 @@
 
 .modal {
     width: 500px;
-    height: 300px;
+    height: 250px;
     border-radius: 5px;
     background-color: whitesmoke;
     padding: 0.5rem;
@@ -20,10 +20,9 @@
     font-size: 1.5rem;
     align-items: center;
 }
+
 .modal-title {
     display: flex;
-    justify-content: space-between;
-    flex-direction: row-reverse;
     align-items: center;
     padding: 0 0.5rem;
     width: 90%;
@@ -38,17 +37,23 @@
 }
 
 .close-modal-button {
-    color:black;
+    color: black;
     background-color: white;
     font-size: 24px;
     font-weight: bold;
     border: none;
- }
+}
 
- .modal-input {
-     display: flex;
-     flex-direction: column;
-     width: 60%;
-     justify-content: center;
- }
- 
+.modal-input {
+    display: flex;
+    flex-direction: column;
+    width: 60%;
+    justify-content: center;
+}
+
+.modal-footer {
+    display: flex;
+    margin: 1rem;
+    justify-content: flex-end;
+    width: 90%;
+}
