+++ my-app/src/components/titles/index.jsx
@@ -1,12 +1,10 @@
 import React from 'react';
 import './style.css';
 import TittleText from './text/text'
-import TittleIcon from './icon/tittleIcon'
 
-const Tittle = ({color, text, icon}) => {
+const Tittle = ({color, text}) => {
     return (
         <header style={{backgroundColor:color}}className="header">
-            <TittleIcon iconLink={icon} className="icon"/>
             <TittleText text={text}/>
         </header>
     )
