+++ my-app/src/components/titles/style.css
@@ -3,6 +3,9 @@
         color: white;
         align-items: center;
         border-radius: 10px;
+        padding: 2rem;
+        width: 95%;
+        height: 10vh;
     }
     
     .icon {
