+++ my-app/src/views/settings/settings.css
@@ -1,0 +1,6 @@
+.main {
+    display: flex;
+    justify-content: center;
+    padding: 1rem;
+    flex-direction: column;
+}
